Added profiler, performance up by 200%
I was playing around with this again, and noticed that the old version
of it performed faster than the new one. This doesn't make sense since I
optimized the check order of my collision detection in the new version.

I found out that it was because I was using deepcopy() instead of copy()
in my checkOnBlock() function. I simply changed it to copy and now the
FPS (when not throttled) is improved by up to 3x, nearly twice as fast
as the old version.

I also fixed a weird derp with my level loading. Nothing big, but
another level.lvl needs to be in the same folder for oldmain.py to work.