Separated idle animations, changed level
I made it so the player can have an idle animation separate from the
running ones. The level layout has also changed. I faked a way to make
the top of the level look "open". Really, the camera just stops before
it reaches the boundary. Other small things include the ability to
change FPS limit on the fly and OSD is disabled by default.